% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDaysOfWeek.R
\name{getDaysOfWeek}
\alias{getDaysOfWeek}
\title{Esta funcion devuelve un vector con los nombres de los dias de la semana}
\usage{
getDaysOfWeek()
}
\description{
Esta funcion devuelve un vector con los nombres de los dias de la semana
}
\examples{
getDaysOfWeek()
}
