% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSequence.R
\name{getSequence}
\alias{getSequence}
\title{Esta funcion devuelve un vector con una secuencia segun los parametros recibidos}
\usage{
getSequence(desde = 1, hasta = 10, paso = 1)
}
\description{
Parámetros: desde, hasta, paso
Si no se envian parametros devolvera la secuencia desde el 1 hasta el 10, con paso 1
}
\examples{
getSequence()
getSequence(1, 20, 2)
}
